+++ src/main/java/com/meta/ponkids/domain/user/dto/UserSaveDto.java
@@ -50,7 +50,7 @@ public class UserSaveDto {
     
     private String detailAdr;       // 상세 주소
     
-    private Integer atchFileSn;      // 첨부파일 일련번호
+    private Long atchFileSn;      // 첨부파일 일련번호
     
     @NotNull
     private String mngrYn;          // 관리자 여부
@@ -72,7 +72,7 @@ public class UserSaveDto {
     
     // builder 생성
     @Builder
-    public UserSaveDto( String userId, String password, String userNm, String gender, String brdtDate, String telNo, String resideArea, String zip, String rdnmAdr, String detailAdr, Integer atchFileSn, String mngrYn, String mngrConfmYn, String confmerId, String confmerIp, LocalDateTime confmDt, String cntnSns, String registerIp, String delYn ) {
+    public UserSaveDto( String userId, String password, String userNm, String gender, String brdtDate, String telNo, String resideArea, String zip, String rdnmAdr, String detailAdr, Long atchFileSn, String mngrYn, String mngrConfmYn, String confmerId, String confmerIp, LocalDateTime confmDt, String cntnSns, String registerIp, String delYn ) {
         this.userId = userId;
         this.password = password;
         this.userNm = userNm;
